+++ src/gwt/src/org/rstudio/studio/client/rsconnect/events/EnableRStudioConnectUIEvent.java
@@ -30,16 +30,11 @@ public class EnableRStudioConnectUIEvent
       public final native boolean getConnectEnable() /*-{
          return this.connectEnabled;
       }-*/;
-
-      public final native boolean getCloudEnable() /*-{
-          return this.cloudEnabled;
-      }-*/;
    }
 
    public interface Handler extends EventHandler
    {
       void onEnableRStudioConnectUI(EnableRStudioConnectUIEvent event);
-      void onEnablePositCloudUI(EnableRStudioConnectUIEvent event);
    }
 
    public static final GwtEvent.Type<EnableRStudioConnectUIEvent.Handler> TYPE = new GwtEvent.Type<>();
@@ -54,16 +49,10 @@ public class EnableRStudioConnectUIEvent
       return data_.getConnectEnable();
    }
 
-   public boolean getCloudEnable()
-   {
-      return data_.getCloudEnable();
-   }
-   
    @Override
    protected void dispatch(EnableRStudioConnectUIEvent.Handler handler)
    {
       handler.onEnableRStudioConnectUI(this);
-      handler.onEnablePositCloudUI(this);
    }
 
    @Override