+++ src/gwt/src/org/rstudio/studio/client/workbench/prefs/model/UserPrefsAccessor.java
@@ -1227,7 +1227,7 @@ public class UserPrefsAccessor extends Prefs
          "console_line_length_limit",
          _constants.consoleLineLengthLimitTitle(), 
          _constants.consoleLineLengthLimitDescription(), 
-         1000);
+         2000);
    }
 
    /**
@@ -1239,7 +1239,7 @@ public class UserPrefsAccessor extends Prefs
          "console_max_lines",
          _constants.consoleMaxLinesTitle(), 
          _constants.consoleMaxLinesDescription(), 
-         1000);
+         10000);
    }
 
    /**
@@ -1842,18 +1842,6 @@ public class UserPrefsAccessor extends Prefs
          _constants.showPublishDiagnosticsTitle(), 
          _constants.showPublishDiagnosticsDescription(), 
          false);
-   }
-
-   /**
-    * Whether to show UI for publishing content to Posit Cloud.
-    */
-   public PrefValue<Boolean> enableCloudPublishUi()
-   {
-      return bool(
-         "enable_cloud_publish_ui",
-         _constants.enableCloudPublishUiTitle(), 
-         _constants.enableCloudPublishUiDescription(), 
-         true);
    }
 
    /**
@@ -4082,8 +4070,6 @@ public class UserPrefsAccessor extends Prefs
          rmdViewerType().setValue(layer, source.getString("rmd_viewer_type"));
       if (source.hasKey("show_publish_diagnostics"))
          showPublishDiagnostics().setValue(layer, source.getBool("show_publish_diagnostics"));
-      if (source.hasKey("enable_cloud_publish_ui"))
-         enableCloudPublishUi().setValue(layer, source.getBool("enable_cloud_publish_ui"));
       if (source.hasKey("publish_check_certificates"))
          publishCheckCertificates().setValue(layer, source.getBool("publish_check_certificates"));
       if (source.hasKey("use_publish_ca_bundle"))
@@ -4491,7 +4477,6 @@ public class UserPrefsAccessor extends Prefs
       prefs.add(rmdPreferredTemplatePath());
       prefs.add(rmdViewerType());
       prefs.add(showPublishDiagnostics());
-      prefs.add(enableCloudPublishUi());
       prefs.add(publishCheckCertificates());
       prefs.add(usePublishCaBundle());
       prefs.add(publishCaBundle());